requirements.txt
ansicolors==1.0.2
cycler==0.10.0
h5py==2.7.0
Keras==2.0.4
matplotlib==2.0.1
nltk==3.2.4
numpy==1.12.1
pandas==0.20.1
protobuf==3.3.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
PyYAML==3.12
scikit-learn==0.18.1
scipy==0.19.0
six==1.10.0
sklearn==0.0
tensorflow==1.1.0
Theano==0.9.0
Werkzeug==0.12.1
